chore(ci): Detox runner split refactor
