+++ suite-native/app/package.json
@@ -18,9 +18,9 @@
         "eas-build-on-success": "./eas-post-success.sh",
         "prebuild:clean": "expo prebuild --clean",
         "build:e2e": "npx detox build --configuration",
-        "test:e2e": "node ./scripts/cleanArtifacts.js && npx tsx e2e/trezorDetoxRunner/detoxRunner.ts",
-        "test:e2e:android": "node ./scripts/cleanArtifacts.js && npx tsx e2e/trezorDetoxRunner/detoxRunner.ts --config=e2e/trezorDetoxRunner/config/runner-android.config.js",
-        "test:e2e:ios": "node ./scripts/cleanArtifacts.js && npx tsx e2e/trezorDetoxRunner/detoxRunner.ts --config=e2e/trezorDetoxRunner/config/runner-ios.config.js",
+        "test:e2e": "node ./scripts/cleanArtifacts.js && npx tsx e2e/trezorDetoxRunner/index.ts",
+        "test:e2e:android": "node ./scripts/cleanArtifacts.js && npx tsx e2e/trezorDetoxRunner/index.ts --config=e2e/trezorDetoxRunner/config/runner-android.config.js",
+        "test:e2e:ios": "node ./scripts/cleanArtifacts.js && npx tsx e2e/trezorDetoxRunner/index.ts --config=e2e/trezorDetoxRunner/config/runner-ios.config.js",
         "start:e2e": "NODE_ENV=test RN_SRC_EXT=e2e.ts,e2e.tsx expo start --dev-client",
         "reverse-ports": "adb reverse tcp:8081 tcp:8081 && adb reverse tcp:21328 tcp:21328 && adb reverse tcp:19121 tcp:19121",
         "github:report:manual": "npx jest --config=./e2e/jest.config.reporter.js"
